+++ ameriflux_pipeline/main.py
@@ -226,7 +226,6 @@ def pyfluxpro_l2_ameriflux_processing(pyfluxpro_ameriflux_labels, l2_mainstem, l
 if __name__ == '__main__':
     # run eddypro preprocessing and formatting
     eddypro_formatted_met_file, file_meta_data_file = eddypro_preprocessing()
-
     # run eddypro
     run_eddypro(eddypro_formatted_met_file)
 
@@ -251,6 +250,7 @@ if __name__ == '__main__':
         print(cfg.PYFLUXPRO_INPUT_SHEET, "path does not exist")
 
     # run ameriflux formatting of pyfluxpro L1 control file
+<<<<<<< HEAD
     pyfluxpro_ameriflux_labels = pyfluxpro_l1_ameriflux_processing(cfg.PYFLUXPRO_INPUT_AMERIFLUX, cfg.L1_MAINSTEM,
                                                                    cfg.L1_AMERIFLUX_ONLY, cfg.L1_AMERIFLUX_MAINSTEM_KEY,
                                                                    file_meta_data_file, cfg.INPUT_SOIL_KEY,
@@ -261,3 +261,9 @@ if __name__ == '__main__':
     pyfluxpro_l2_ameriflux_processing(pyfluxpro_ameriflux_labels, l2_mainstem, l2_ameriflux_only,
                                       l1_run_output, l2_run_output, l2_ameriflux_output)
 
+=======
+    pyfluxpro_l1_ameriflux_processing(cfg.PYFLUXPRO_INPUT_AMERIFLUX, cfg.L1_MAINSTEM_INPUT, cfg.L1_AMERIFLUX_ONLY_INPUT,
+                                      cfg.L1_AMERIFLUX_MAINSTEM_KEY, file_meta_data_file, cfg.INPUT_SOIL_KEY,
+                                      cfg.L1_AMERIFLUX_RUN_OUTPUT, cfg.L1_AMERIFLUX,
+                                      cfg.AMERIFLUX_VARIABLE_USER_CONFIRMATION, cfg.L1_AMERIFLUX_ERRORING_VARIABLES_KEY)
+>>>>>>> 95593c4b7a39340fe8487920d1db77b60831370e