+++ CHANGELOG.md
@@ -30,6 +30,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - L1 processing is made more efficient [#100](https://github.com/ncsa/ameriflux-pipeline/issues/100)
 - Added user confirmation to replace certain PyFluxPro variable names to AmeriFlux names [#105](https://github.com/ncsa/ameriflux-pipeline/issues/105)
 - Check formatting of L1 inputs [#109](https://github.com/ncsa/ameriflux-pipeline/issues/109)
+- Updating units for Ameriflux only variables [#115](https://github.com/ncsa/ameriflux-pipeline/issues/115)
+- Check formatting of L1 inputs [#109](https://github.com/ncsa/ameriflux-pipeline/issues/109)
 
 ### Fixed
 - Python packaging of the codes so it can run in command prompt. [#70](https://github.com/ncsa/ameriflux-pipeline/issues/70)